BulkButton Fix
The bulk action was missing a true command, so bulk active would never be set to true. I fixed this by setting bulkActive to true in the else clause and adding an if clause in deleteAction to insure that the bulk button doesn't get off sequence when the deleteAction clears the bulk button's state.